pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dtvanshul-telegram-cleaner"
version = "0.1.1"
description = "Powerful Telegram channel cleaner CLI with userbot control, batch deletion, and progress tracking."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "dtvanshulll" },
]
keywords = ["telegram", "userbot", "automation", "cleaner"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Communications :: Chat",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities",
]
dependencies = [
  "colorama>=0.4.6",
  "python-dotenv>=1.0.1",
  "telebridge",
  "tqdm>=4.62.3",
]

[project.optional-dependencies]
publish = [
  "build>=1.2.2",
  "twine>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/dtvanshulll/Telegram_cleaner"
Repository = "https://github.com/dtvanshulll/Telegram_cleaner"
Issues = "https://github.com/dtvanshulll/Telegram_cleaner/issues"

[project.scripts]
telegramcleaner = "telegramcleaner.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["telegramcleaner*"]
